Rename package: news_sources -> newssources
