+++ app/src/main/java/com/voltek/newsfeed/dagger/AppComponent.kt
@@ -2,7 +2,7 @@ package com.voltek.newsfeed.dagger
 
 import com.voltek.newsfeed.presentation.ui.details.DetailsFragment
 import com.voltek.newsfeed.presentation.ui.list.ListFragment
-import com.voltek.newsfeed.presentation.ui.news_sources.NewsSourcesFragment
+import com.voltek.newsfeed.presentation.ui.newssources.NewsSourcesFragment
 import com.voltek.newsfeed.presentation.ui.splash.SplashActivity
 import dagger.Component
 import javax.inject.Singleton
