+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/newssources/NewsSourcesActivity.kt
@@ -1,4 +1,4 @@
-package com.voltek.newsfeed.presentation.ui.news_sources
+package com.voltek.newsfeed.presentation.ui.newssources
 
 import android.os.Bundle
 import android.support.design.widget.Snackbar
