+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/newssources/NewsSourcesAdapter.kt
@@ -1,4 +1,4 @@
-package com.voltek.newsfeed.presentation.ui.news_sources
+package com.voltek.newsfeed.presentation.ui.newssources
 
 import android.content.Context
 import android.support.annotation.DrawableRes
